% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_census.R
\name{read_census_keieitai}
\alias{read_census_keieitai}
\title{Read census keieitai file}
\usage{
read_census_keieitai(path, data.table = TRUE)
}
\arguments{
\item{path}{file}

\item{data.table}{Return data class.}
}
\description{
Read census keieitai file
}
