% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machimura.R
\name{view_machimura}
\alias{view_machimura}
\title{View Waga Machi Waga Mura page}
\usage{
view_machimura(city_code, type)
}
\arguments{
\item{city_code}{city code}

\item{type}{"basic" or "detail"}
}
\description{
View Waga Machi Waga Mura page
}
\examples{
\dontrun{
build_machimura("08201", type = "base")
}
}
